feat(generator): [NFD] Update generator
