+++ buildSrc/build.gradle.kts
@@ -1,9 +1,11 @@
 plugins { `kotlin-dsl` }
 
-dependencies {
-    implementation("org.jetbrains.kotlin.jvm:org.jetbrains.kotlin.jvm.gradle.plugin:2.0.20")
-}
+val kotlinPluginVersion: String by rootProject
 
 repositories {
     mavenCentral()
+}
+
+dependencies {
+    implementation("org.jetbrains.kotlin.jvm:org.jetbrains.kotlin.jvm.gradle.plugin:$kotlinPluginVersion")
 }
