+++ buildSrc/gradle.properties
@@ -1,0 +1,1 @@
+kotlinPluginVersion=2.1.0
