+++ gradle.properties
@@ -10,9 +10,13 @@ projLicenseYear=2022-2024
 jdkVersion=23
 jdkEnablePreview=true
 #jdkEarlyAccessDoc=jdk23
-kotlinTargetJdkVersion=21
+kotlinTargetJdkVersion=22
 
-overrunMarshalVersion=0.1.0-alpha.37-jdk23
+overrunMarshalVersion=0.1.0-alpha.38-jdk23
 overrunPlatformVersion=1.0.0
+javaPoetVersion=0.5.0
+jetbrainsAnnotationsVersion=26.0.1
+jmhVersion=1.37
 jomlVersion=1.10.8
-junitVersion=5.11.0
+junitVersion=5.11.3
+timerVersion=0.3.0
