+++ modules/overrungl.glfw/src/main/java/overrungl/glfw/GLFW.java
@@ -22,7 +22,6 @@ import overrun.marshal.DirectAccess;
 import overrun.marshal.Downcall;
 import overrun.marshal.Unmarshal;
 import overrun.marshal.gen.*;
-import overrun.marshal.gen.processor.ProcessorType.BoolConvert;
 import overrungl.internal.RuntimeHelper;
 import overrungl.util.value.Pair;
 import overrungl.util.value.Quad;
@@ -1235,7 +1234,7 @@ public interface GLFW extends DirectAccess {
      * @see #initAllocator
      * @see #terminate
      */
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("glfwInit")
     boolean init();
 
@@ -1527,7 +1526,7 @@ public interface GLFW extends DirectAccess {
      * @glfw.thread_safety This function may be called from any thread.
      * @see #getPlatform
      */
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("glfwPlatformSupported")
     boolean platformSupported(int platform);
 
@@ -2146,7 +2145,7 @@ public interface GLFW extends DirectAccess {
      * @see #windowHint(int, int)
      */
     @Entrypoint("glfwWindowHint")
-    void windowHint(int hint, @Convert(BoolConvert.INT) boolean value);
+    void windowHint(int hint, @Convert(AsBool.INT) boolean value);
 
     /**
      * Sets the specified window hint to the desired value.
@@ -2393,7 +2392,7 @@ public interface GLFW extends DirectAccess {
      * @glfw.thread_safety This function may be called from any thread.  Access is not
      * synchronized.
      */
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("glfwWindowShouldClose")
     boolean windowShouldClose(MemorySegment window);
 
@@ -2411,7 +2410,7 @@ public interface GLFW extends DirectAccess {
      * synchronized.
      */
     @Entrypoint("glfwSetWindowShouldClose")
-    void setWindowShouldClose(MemorySegment window, @Convert(BoolConvert.INT) boolean value);
+    void setWindowShouldClose(MemorySegment window, @Convert(AsBool.INT) boolean value);
 
     /**
      * Returns the title of the specified window.
@@ -3294,7 +3293,7 @@ public interface GLFW extends DirectAccess {
      * @see #getWindowAttrib(MemorySegment, int) getWindowAttrib
      */
     @Entrypoint("glfwSetWindowAttrib")
-    void setWindowAttrib(MemorySegment window, int attrib, @Convert(BoolConvert.INT) boolean value);
+    void setWindowAttrib(MemorySegment window, int attrib, @Convert(AsBool.INT) boolean value);
 
     /**
      * Sets the user pointer of the specified window.
@@ -3899,7 +3898,7 @@ public interface GLFW extends DirectAccess {
      * @glfw.thread_safety This function must only be called from the main thread.
      * @see #setInputMode(MemorySegment, int, int) setInputMode
      */
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("glfwRawMouseMotionSupported")
     boolean rawMouseMotionSupported();
 
@@ -4591,7 +4590,7 @@ public interface GLFW extends DirectAccess {
      * {@link #INVALID_ENUM} and {@link #PLATFORM_ERROR}.
      * @glfw.thread_safety This function must only be called from the main thread.
      */
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("glfwJoystickPresent")
     boolean joystickPresent(int jid);
 
@@ -4899,7 +4898,7 @@ public interface GLFW extends DirectAccess {
      * @glfw.thread_safety This function must only be called from the main thread.
      * @see #ngetGamepadState(int, MemorySegment) getGamepadState
      */
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("glfwJoystickIsGamepad")
     boolean joystickIsGamepad(int jid);
 
@@ -4970,7 +4969,7 @@ public interface GLFW extends DirectAccess {
      * @see #joystickIsGamepad(int) joystickIsGamepad
      * @see #ngetGamepadName(int) getGamepadName
      */
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("glfwUpdateGamepadMappings")
     boolean nupdateGamepadMappings(MemorySegment string);
 
@@ -4982,7 +4981,7 @@ public interface GLFW extends DirectAccess {
      * <a href="https://www.glfw.org/docs/latest/intro_guide.html#error_handling">error</a> occurred.
      * @see #nupdateGamepadMappings(MemorySegment) nupdateGamepadMappings
      */
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("glfwUpdateGamepadMappings")
     boolean updateGamepadMappings(String string);
 
@@ -5053,7 +5052,7 @@ public interface GLFW extends DirectAccess {
      * @see #nupdateGamepadMappings(MemorySegment) updateGamepadMappings
      * @see #joystickIsGamepad(int) joystickIsGamepad
      */
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("glfwGetGamepadState")
     boolean ngetGamepadState(int jid, MemorySegment state);
 
@@ -5067,7 +5066,7 @@ public interface GLFW extends DirectAccess {
      * occurred.
      * @see #ngetGamepadState(int, MemorySegment) ngetGamepadState
      */
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("glfwGetGamepadState")
     boolean getGamepadState(int jid, GLFWGamepadState state);
 
@@ -5414,7 +5413,7 @@ public interface GLFW extends DirectAccess {
      * @glfw.thread_safety This function may be called from any thread.
      * @see #ngetProcAddress
      */
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("glfwExtensionSupported")
     boolean nextensionSupported(MemorySegment extension);
 
@@ -5426,7 +5425,7 @@ public interface GLFW extends DirectAccess {
      * otherwise.
      * @see #nextensionSupported(MemorySegment) nextensionSupported
      */
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("glfwExtensionSupported")
     boolean extensionSupported(@StrCharset("US-ASCII") String extension);
 
@@ -5494,7 +5493,7 @@ public interface GLFW extends DirectAccess {
      * @glfw.errors Possible errors include {@link #NOT_INITIALIZED}.
      * @glfw.thread_safety This function may be called from any thread.
      */
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("glfwVulkanSupported")
     boolean vulkanSupported();
