+++ modules/overrungl.glfw/src/main/java/overrungl/glfw/GLFWVulkan.java
@@ -18,10 +18,10 @@ package overrungl.glfw;
 
 import overrun.marshal.DirectAccess;
 import overrun.marshal.Downcall;
+import overrun.marshal.gen.AsBool;
 import overrun.marshal.gen.Convert;
 import overrun.marshal.gen.Entrypoint;
 import overrun.marshal.gen.Ref;
-import overrun.marshal.gen.processor.ProcessorType.BoolConvert;
 import overrungl.NativeType;
 
 import java.lang.foreign.MemorySegment;
@@ -149,7 +149,7 @@ public interface GLFWVulkan extends DirectAccess {
      * @glfw.thread_safety This function may be called from any thread.  For
      * synchronization details of Vulkan objects, see the Vulkan specification.
      */
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("glfwGetPhysicalDevicePresentationSupport")
     boolean getPhysicalDevicePresentationSupport(MemorySegment instance, MemorySegment device, int queueFamily);
