+++ modules/overrungl.nfd/src/main/java/overrungl/nfd/NFDEnumerator.java
@@ -16,15 +16,13 @@
 
 package overrungl.nfd;
 
+import io.github.overrun.memstack.MemoryStack;
 import org.jetbrains.annotations.NotNull;
-import overrun.marshal.struct.Struct;
-import overrun.marshal.struct.StructAllocator;
+import overrun.marshal.Unmarshal;
 import overrungl.util.value.Tuple2;
 
-import java.lang.foreign.MemoryLayout;
 import java.lang.foreign.MemorySegment;
 import java.lang.foreign.SegmentAllocator;
-import java.lang.invoke.MethodHandles;
 import java.util.Iterator;
 import java.util.NoSuchElementException;
 
@@ -32,22 +30,14 @@ import static java.lang.foreign.ValueLayout.ADDRESS;
 
 /**
  * A wrapper of NFD path set enumerator.
+ * <p>
+ * Users are responsible to close the instance after the iteration finished.
  *
  * @author squid233
  * @see NFD#pathSetGetEnum(MemorySegment, MemorySegment)
  * @since 0.1.0
  */
 public final class NFDEnumerator implements Iterable<String>, AutoCloseable {
-    private interface Segment extends Struct<Segment> {
-        /**
-         * The allocator
-         */
-        StructAllocator<Segment> OF = new StructAllocator<>(
-            MethodHandles.lookup(),
-            MemoryLayout.structLayout(ADDRESS.withName("ptr"))
-        );
-    }
-
     private static final Iterator<String> EMPTY_ITERATOR = new Iterator<>() {
         @Override
         public boolean hasNext() {
@@ -59,30 +49,10 @@ public final class NFDEnumerator implements Iterable<String>, AutoCloseable {
             throw new NoSuchElementException();
         }
     };
-    private final NFD nfd = NFD.INSTANCE;
-    private final Kind kind;
-    private final Segment segment;
+    private final MemorySegment segment;
 
-    private NFDEnumerator(Kind kind, Segment segment) {
-        this.kind = kind;
+    private NFDEnumerator(MemorySegment segment) {
         this.segment = segment;
-    }
-
-    /**
-     * The kind of the enumerator.
-     *
-     * @author squid233
-     * @since 0.1.0
-     */
-    public enum Kind {
-        /**
-         * Native (UTF-16LE for Windows, UTF-8 for others)
-         */
-        N,
-        /**
-         * UTF-8
-         */
-        U8
     }
 
     /**
@@ -107,24 +77,17 @@ public final class NFDEnumerator implements Iterable<String>, AutoCloseable {
             if (curr == null) {
                 throw new NoSuchElementException();
             }
-            String[] s = new String[1];
-            final int result = switch (kind) {
-                case N -> nfd.pathSetEnumNextN(segment.segment(), s);
-                case U8 -> nfd.pathSetEnumNextU8(segment.segment(), s);
-            };
-            if (result == NFD.ERROR) throw errorIterating(nfd);
-            nextPath = s[0];
+            try (MemoryStack stack = MemoryStack.pushLocal()) {
+                MemorySegment outPath = stack.allocate(ADDRESS);
+                int result = NFD.pathSetEnumNextN(segment, outPath);
+                if (result == NFD.ERROR) throw errorIterating();
+                nextPath = Unmarshal.unmarshalStringPointer(outPath, NFDInternal.nfdCharset);
+                if (nextPath != null) {
+                    NFD.pathSetFreePathN(outPath.get(ADDRESS, 0));
+                }
+            }
             return curr;
         }
-    }
-
-    private static Tuple2.OfObjInt<NFDEnumerator> fromPathSet(Kind kind, SegmentAllocator allocator, MemorySegment pathSet) {
-        final MemorySegment seg = allocator.allocate(ADDRESS);
-        final int result = NFD.INSTANCE.pathSetGetEnum(pathSet, seg);
-        return new Tuple2.OfObjInt<>(result == NFD.OKAY ?
-                new NFDEnumerator(kind, Segment.OF.of(seg)) :
-                null,
-            result);
     }
 
     /**
@@ -135,45 +98,43 @@ public final class NFDEnumerator implements Iterable<String>, AutoCloseable {
      * @return the result and the enumerator.
      */
     public static Tuple2.OfObjInt<NFDEnumerator> fromPathSetN(SegmentAllocator allocator, MemorySegment pathSet) {
-        return fromPathSet(Kind.N, allocator, pathSet);
+        MemorySegment struct = allocator.allocate(ADDRESS);
+        int result = NFD.pathSetGetEnum(pathSet, struct);
+        return new Tuple2.OfObjInt<>(result == NFD.OKAY ?
+            new NFDEnumerator(struct) :
+            null,
+            result);
     }
 
-    /**
-     * Creates an enumerator from the given path set that is created with {@link NFD#openDialogMultipleU8}.
-     *
-     * @param allocator the allocator of the enumerator.
-     * @param pathSet   the path set.
-     * @return the result and the enumerator.
-     */
-    public static Tuple2.OfObjInt<NFDEnumerator> fromPathSetU8(SegmentAllocator allocator, MemorySegment pathSet) {
-        return fromPathSet(Kind.U8, allocator, pathSet);
-    }
-
-    private static IllegalStateException errorIterating(NFD nfd) {
-        return new IllegalStateException("Error iterating: " + nfd.getError());
+    private static IllegalStateException errorIterating() {
+        return new IllegalStateException("Error iterating: " + NFD.getError());
     }
 
     @NotNull
     @Override
     public Iterator<String> iterator() {
         // TODO: 2023/7/6 Value object
-        String[] s = new String[1];
-        final int result = switch (kind) {
-            case N -> nfd.pathSetEnumNextN(segment.segment(), s);
-            case U8 -> nfd.pathSetEnumNextU8(segment.segment(), s);
-        };
-        final String path = s[0];
-        if (path == null || result != NFD.OKAY) {
-            if (result == NFD.ERROR) {
-                throw errorIterating(nfd);
+        try (MemoryStack stack = MemoryStack.pushLocal()) {
+            MemorySegment outPath = stack.allocate(ADDRESS);
+            int result = NFD.pathSetEnumNextN(segment, outPath);
+            switch (result) {
+                case NFD.OKAY -> {
+                    String path = Unmarshal.unmarshalStringPointer(outPath, NFDInternal.nfdCharset);
+                    if (path != null) {
+                        NFD.pathSetFreePathN(outPath.get(ADDRESS, 0));
+                    }
+                    return new EnumIterator(path);
+                }
+                case NFD.CANCEL -> {
+                    return EMPTY_ITERATOR;
+                }
+                default -> throw errorIterating();
             }
-            return EMPTY_ITERATOR;
         }
-        return new EnumIterator(path);
     }
 
     @Override
     public void close() {
-        nfd.pathSetFreeEnum(segment.segment());
+        NFD.pathSetFreeEnum(segment);
     }
 }
