+++ modules/overrungl.stb/src/main/java/overrungl/stb/STBImageWrite.java
@@ -18,10 +18,10 @@ package overrungl.stb;
 
 import overrun.marshal.DirectAccess;
 import overrun.marshal.Downcall;
+import overrun.marshal.gen.AsBool;
 import overrun.marshal.gen.Convert;
 import overrun.marshal.gen.Entrypoint;
 import overrun.marshal.gen.Skip;
-import overrun.marshal.gen.processor.ProcessorType.BoolConvert;
 
 import java.lang.foreign.Arena;
 import java.lang.foreign.MemorySegment;
@@ -71,88 +71,88 @@ public interface STBImageWrite extends DirectAccess {
         Handles.stbi_write_force_png_filter.set(ValueLayout.JAVA_INT, 0L, filter);
     }
 
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("stbi_write_png")
     boolean npng(MemorySegment filename, int w, int h, int comp, MemorySegment data, int strideInBytes);
 
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("stbi_write_bmp")
     boolean nbmp(MemorySegment filename, int w, int h, int comp, MemorySegment data);
 
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("stbi_write_tga")
     boolean ntga(MemorySegment filename, int w, int h, int comp, MemorySegment data);
 
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("stbi_write_hdr")
     boolean nhdr(MemorySegment filename, int w, int h, int comp, MemorySegment data);
 
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("stbi_write_jpg")
     boolean njpg(MemorySegment filename, int x, int y, int comp, MemorySegment data, int quality);
 
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("stbi_write_png")
     boolean png(String filename, int w, int h, int comp, MemorySegment data, int strideInBytes);
 
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("stbi_write_bmp")
     boolean bmp(String filename, int w, int h, int comp, MemorySegment data);
 
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("stbi_write_tga")
     boolean tga(String filename, int w, int h, int comp, MemorySegment data);
 
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("stbi_write_hdr")
     boolean hdr(SegmentAllocator allocator, String filename, int w, int h, int comp, float[] data);
 
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("stbi_write_jpg")
     boolean jpg(String filename, int x, int y, int comp, MemorySegment data, int quality);
 
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("stbi_write_png_to_func")
     boolean npngToFunc(MemorySegment func, MemorySegment context, int w, int h, int comp, MemorySegment data, int strideInBytes);
 
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("stbi_write_bmp_to_func")
     boolean nbmpToFunc(MemorySegment func, MemorySegment context, int w, int h, int comp, MemorySegment data);
 
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("stbi_write_tga_to_func")
     boolean ntgaToFunc(MemorySegment func, MemorySegment context, int w, int h, int comp, MemorySegment data);
 
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("stbi_write_hdr_to_func")
     boolean nhdrToFunc(MemorySegment func, MemorySegment context, int w, int h, int comp, MemorySegment data);
 
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("stbi_write_jpg_to_func")
     boolean njpgToFunc(MemorySegment func, MemorySegment context, int x, int y, int comp, MemorySegment data, int quality);
 
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("stbi_write_png_to_func")
     boolean pngToFunc(Arena arena, STBIWriteFunc func, MemorySegment context, int w, int h, int comp, MemorySegment data, int strideInBytes);
 
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("stbi_write_bmp_to_func")
     boolean bmpToFunc(Arena arena, STBIWriteFunc func, MemorySegment context, int w, int h, int comp, MemorySegment data);
 
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("stbi_write_tga_to_func")
     boolean tgaToFunc(Arena arena, STBIWriteFunc func, MemorySegment context, int w, int h, int comp, MemorySegment data);
 
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("stbi_write_hdr_to_func")
     boolean hdrToFunc(Arena arena, STBIWriteFunc func, MemorySegment context, int w, int h, int comp, float[] data);
 
-    @Convert(BoolConvert.INT)
+    @Convert(AsBool.INT)
     @Entrypoint("stbi_write_jpg_to_func")
     boolean jpgToFunc(Arena arena, STBIWriteFunc func, MemorySegment context, int x, int y, int comp, MemorySegment data, int quality);
 
     @Entrypoint("stbi_flip_vertically_on_write")
-    void flipVerticallyOnWrite(@Convert(BoolConvert.INT) boolean flip);
+    void flipVerticallyOnWrite(@Convert(AsBool.INT) boolean flip);
 
     @Entrypoint("stbi_write_png_to_mem")
     MemorySegment npngToMem(MemorySegment pixels, int strideInBytes, int x, int y, int n, MemorySegment outLen);
