+++ modules/samples/build.gradle.kts
@@ -8,15 +8,17 @@ overrunglModule {
     artifactName = "overrungl-samples"
 }
 
-val junitVersion: String by project
+val jmhVersion: String by rootProject
+val junitVersion: String by rootProject
+val timerVersion: String by rootProject
 
 dependencies {
     Artifact.values().forEach {
         implementation(project(it.subprojectName))
     }
-    implementation("io.github.over-run:timer:0.3.0")
-    jmh("org.openjdk.jmh:jmh-core:1.37")
-    jmhAnnotationProcessor("org.openjdk.jmh:jmh-generator-annprocess:1.37")
+    implementation("io.github.over-run:timer:$timerVersion")
+    jmh("org.openjdk.jmh:jmh-core:$jmhVersion")
+    jmhAnnotationProcessor("org.openjdk.jmh:jmh-generator-annprocess:$jmhVersion")
     testImplementation("org.junit.jupiter:junit-jupiter:$junitVersion")
 }
