+++ modules/samples/src/main/java/overrungl/demo/nfd/NFDTest.java
@@ -17,9 +17,9 @@
 package overrungl.demo.nfd;
 
 import io.github.overrun.memstack.MemoryStack;
-import overrungl.nfd.NFD;
 import overrungl.nfd.NFDEnumerator;
 import overrungl.nfd.NFDNFilterItem;
+import overrungl.nfd.NFD;
 
 import java.lang.foreign.Arena;
 import java.lang.foreign.MemorySegment;
@@ -31,14 +31,12 @@ import java.util.Map;
  * @since 0.1.0
  */
 public final class NFDTest {
-    private static final NFD nfd = NFD.INSTANCE;
-
     private static void openDialog() {
         System.out.println("===== openDialog =====");
         // initialize NFD
         // either call NFD::init at the start of your program and NFD::quit at the end of your program,
         // or before/after every time you want to show a file dialog.
-        nfd.init();
+        NFD.init();
 
         try (MemoryStack stack = MemoryStack.pushLocal()) {
             String[] outPath = new String[1];
@@ -49,17 +47,17 @@ public final class NFDTest {
                 Map.entry("Image file", "png,jpg"));
 
             // show the dialog
-            final int result = nfd.openDialogN(outPath, filterItem, null);
+            final int result = NFD.openDialogN(outPath, filterItem, null);
 
             switch (result) {
-                case NFD.ERROR -> System.err.println("Error: " + nfd.getError());
+                case NFD.ERROR -> System.err.println("Error: " + NFD.getError());
                 case NFD.OKAY -> System.out.println("Success! " + outPath[0]);
                 case NFD.CANCEL -> System.out.println("User pressed cancel.");
             }
         }
 
         // Quit NFD
-        nfd.quit();
+        NFD.quit();
     }
 
     private static void openDialogMultiple() {
@@ -67,7 +65,7 @@ public final class NFDTest {
         // initialize NFD
         // either call NFD::init at the start of your program and NFD::quit at the end of your program,
         // or before/after every time you want to show a file dialog.
-        nfd.init();
+        NFD.init();
 
         try (MemoryStack stack = MemoryStack.pushLocal()) {
             MemorySegment pOutPaths = stack.allocate(ValueLayout.ADDRESS);
@@ -79,28 +77,30 @@ public final class NFDTest {
                 Map.entry("Image file", "png,jpg"));
 
             // show the dialog
-            final int result = nfd.openDialogMultipleN(pOutPaths, filterItem, null);
-            MemorySegment outPaths = pOutPaths.get(ValueLayout.ADDRESS, 0).reinterpret(Arena.global(), nfd::pathSetFreePathN);
+            final int result = NFD.openDialogMultipleN(pOutPaths, filterItem, null);
+            MemorySegment outPaths = pOutPaths.get(ValueLayout.ADDRESS, 0);
 
             switch (result) {
-                case NFD.ERROR -> System.err.println("Error: " + nfd.getError());
+                case NFD.ERROR -> System.err.println("Error: " + NFD.getError());
                 case NFD.OKAY -> {
                     System.out.println("Success!");
 
-                    for (long i = 0, numPaths = nfd.pathSetGetCount(outPaths).y(); i < numPaths; i++) {
-                        nfd.pathSetGetPathN(outPaths, i, outPath);
+                    long[] pNumPaths = new long[1];
+                    NFD.pathSetGetCount(outPaths, pNumPaths);
+                    for (long i = 0, numPaths = pNumPaths[0]; i < numPaths; i++) {
+                        NFD.pathSetGetPathN(outPaths, i, outPath);
                         System.out.println("Path " + i + ": " + outPath[0]);
                     }
 
                     // remember to free the path-set memory (since NFDResult::OKAY is returned)
-                    nfd.pathSetFree(outPaths);
+                    NFD.pathSetFree(outPaths);
                 }
                 case NFD.CANCEL -> System.out.println("User pressed cancel.");
             }
         }
 
         // Quit NFD
-        nfd.quit();
+        NFD.quit();
     }
 
     private static void openDialogMultipleEnum() {
@@ -108,7 +108,7 @@ public final class NFDTest {
         // initialize NFD
         // either call NFD::init at the start of your program and NFD::quit at the end of your program,
         // or before/after every time you want to show a file dialog.
-        nfd.init();
+        NFD.init();
 
         try (MemoryStack stack = MemoryStack.pushLocal()) {
             MemorySegment pOutPaths = stack.allocate(ValueLayout.ADDRESS);
@@ -119,11 +119,11 @@ public final class NFDTest {
                 Map.entry("Image file", "png,jpg"));
 
             // show the dialog
-            final int result = nfd.openDialogMultipleN(pOutPaths, filterItem, null);
+            final int result = NFD.openDialogMultipleN(pOutPaths, filterItem, null);
             MemorySegment outPaths = pOutPaths.get(ValueLayout.ADDRESS, 0);
 
             switch (result) {
-                case NFD.ERROR -> System.err.println("Error: " + nfd.getError());
+                case NFD.ERROR -> System.err.println("Error: " + NFD.getError());
                 case NFD.OKAY -> {
                     System.out.println("Success!");
 
@@ -136,14 +136,14 @@ public final class NFDTest {
                     }
 
                     // remember to free the path-set memory (since NFDResult::OKAY is returned)
-                    nfd.pathSetFree(outPaths);
+                    NFD.pathSetFree(outPaths);
                 }
                 case NFD.CANCEL -> System.out.println("User pressed cancel.");
             }
         }
 
         // Quit NFD
-        nfd.quit();
+        NFD.quit();
     }
 
     private static void pickFolder() {
@@ -151,20 +151,20 @@ public final class NFDTest {
         // initialize NFD
         // either call NFD::init at the start of your program and NFD::quit at the end of your program,
         // or before/after every time you want to show a file dialog.
-        nfd.init();
+        NFD.init();
 
         String[] outPath = new String[1];
 
         // show the dialog
-        final int result = nfd.pickFolderN(outPath, null);
+        final int result = NFD.pickFolderN(outPath, null);
         switch (result) {
-            case NFD.ERROR -> System.err.println("Error: " + nfd.getError());
+            case NFD.ERROR -> System.err.println("Error: " + NFD.getError());
             case NFD.OKAY -> System.out.println("Success! " + outPath[0]);
             case NFD.CANCEL -> System.out.println("User pressed cancel.");
         }
 
         // Quit NFD
-        nfd.quit();
+        NFD.quit();
     }
 
     private static void saveDialog() {
@@ -172,7 +172,7 @@ public final class NFDTest {
         // initialize NFD
         // either call NFD::init at the start of your program and NFD::quit at the end of your program,
         // or before/after every time you want to show a file dialog.
-        nfd.init();
+        NFD.init();
 
         try (MemoryStack stack = MemoryStack.pushLocal()) {
             String[] savePath = new String[1];
@@ -183,16 +183,16 @@ public final class NFDTest {
                 Map.entry("Image file", "png,jpg"));
 
             // show the dialog
-            final int result = nfd.saveDialogN(savePath, filterItem, null, "Untitled.java");
+            final int result = NFD.saveDialogN(savePath, filterItem, null, "Untitled.java");
             switch (result) {
-                case NFD.ERROR -> System.err.println("Error: " + nfd.getError());
+                case NFD.ERROR -> System.err.println("Error: " + NFD.getError());
                 case NFD.OKAY -> System.out.println("Success! " + savePath[0]);
                 case NFD.CANCEL -> System.out.println("User pressed cancel.");
             }
         }
 
         // Quit NFD
-        nfd.quit();
+        NFD.quit();
     }
 
     public static void main(String[] args) {
